profile-rest-api/requirements.txt
asgiref==3.2.10
attrs==22.1.0
certifi==2022.9.24
charset-normalizer==2.1.1
coreapi==2.3.3
coreschema==0.0.4
Django==3.1
django-jsonfield-backport==1.0.5
django-messages==0.6.0
django-notifs==4.0.0
django-rest-swagger==2.2.0
djangorestframework==3.12.0
drf-yasg==1.21.4
idna==3.4
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonschema==4.17.0
MarkupSafe==2.1.1
openapi-codec==1.3.2
packaging==21.3
pika==1.3.1
psycopg2==2.9.5
Pygments==2.13.0
pyparsing==3.0.9
pyrsistent==0.19.2
pytz==2022.6
PyYAML==6.0
requests==2.28.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.7
simplejson==3.18.0
sqlparse==0.4.3
swagger-spec-validator==3.0.3
typing-extensions==4.4.0
tzdata==2022.6
uritemplate==4.1.1
urllib3==1.26.12
